Create user sessions and link to clucks

+++ server/controllers/cluckControllers.js
@@ -1,9 +1,12 @@
-const Cluck = require('../models/cluckModel');
-const mongoose = require('mongoose');
+const Cluck = require("../models/cluckModel");
+const User = require("../models/user");
+const mongoose = require("mongoose");
 
 // GET all clucks
 const getAllClucks = async (req, res) => {
-  const clucks = await Cluck.find({}).sort({ createdAt: -1 });
+  const clucks = await Cluck.find({})
+    .populate("user", "userName")
+    .sort({ createdAt: -1 });
 
   res.status(200).json(clucks);
 };
@@ -12,10 +15,10 @@ const getAllClucks = async (req, res) => {
 const getCluck = async (req, res) => {
   const { id } = req.params;
 
-  const cluck = await Cluck.findById(id);
+  const cluck = await Cluck.findById(id).populate("user", "userName");
 
   if (!cluck) {
-    res.status(404).json({ error: 'Cluck not found' });
+    res.status(404).json({ error: "Cluck not found" });
   } else {
     res.status(200).json(cluck);
   }
@@ -24,24 +27,40 @@ const getCluck = async (req, res) => {
 // DELETE a cluck
 const deleteCluck = async (req, res) => {
   const { id } = req.params;
+  const userId = req.userId;
 
+  const cluck = await Cluck.findById(id);
   if (!mongoose.Types.ObjectId.isValid(id)) {
-    return res.status(404).json({ error: 'Cluck not found' });
+    return res.status(404).json({ error: "Cluck not found" });
   }
 
-  const cluck = await Cluck.findByIdAndDelete(id);
+  if (cluck.user._id.toString() !== userId.toString()) {
+    return res
+      .status(403)
+      .json({ message: "You can only delete your own clucks" });
+  }
+
+  await Cluck.findByIdAndDelete(id);
 
   if (!cluck) {
-    return res.status(404).json({ error: 'Cluck not found' });
+    return res.status(404).json({ error: "Cluck not found" });
+  } else {
+    res.status(200).json({ message: "Cluck deleted successfully" });
   }
 };
 
 // POST a new cluck
 const postCluck = async (req, res) => {
-  const { text } = req.body;
-
   try {
-    const cluck = await Cluck.create({ text });
+    const { text } = req.body;
+    const author = await User.findById(req.userId);
+
+    if (!author) {
+      return res.status(400).json({ error: "User not found" });
+    }
+
+    const cluck = await Cluck.create({ text: text, user: author });
+
     res.status(200).json(cluck);
   } catch (error) {
     res.status(400).json({ error: error.message });
@@ -51,18 +70,30 @@ const postCluck = async (req, res) => {
 // PATCH (edit) a cluck
 const editCluck = async (req, res) => {
   const { id } = req.params;
+  const userId = req.userId;
 
-  if (!mongoose.Types.ObjectId.isValid(id)) {
-    return res.status(404).json({ error: 'Cluck not found' });
+  const cluck = await Cluck.findById(id);
+  if (!cluck) {
+    return res.status(404).json({ error: "Cluck not found" });
   }
 
-  const cluck = await Cluck.findOneAndUpdate({ _id: id }, { ...req.body });
+  if (cluck.user._id.toString() !== userId.toString()) {
+    return res
+      .status(403)
+      .json({ message: "You can only edit your own clucks" });
+  }
 
   if (!cluck) {
-    return res.status(404).json({ error: 'Cluck not found' });
+    return res.status(404).json({ error: "Cluck not found" });
   }
 
-  res.status(200).json(cluck);
+  const updatedCluck = await Cluck.findOneAndUpdate(
+    { _id: id },
+    { text: req.body.text },
+    { new: true }
+  );
+
+  res.status(200).json(updatedCluck);
 };
 
 module.exports = {
